test(posthog): split telemetry tests into module
